Death animation script implemented
The same script can be attached to both player and enemy to perform death animation.
Instead of animating it in the update() function, coroutines were used.
Death animation is just a simple rise and fall, so it was implemented using JUST MATH.
The script is attached to Mario and it doesn't do anything for now.
It's execution will be triggered later.