Pit fall death implemented
Mario and enemies die by falling into pits.
Used a trigger collider for this. However I did get an unexpected behaviour where when Mario would fall into the pit without the negative y velocity threshold is met for the collision to happen and he kinda interpolates until his collider touches the trigger collider in the pit. I did fix this by making the pit's trigger collider 2 units below the it, but I think that shouldn't be the case. Welp, for now it does work as intended.
I will later make the DeathAnimation script to be enabled when player or enemy falls into the pit :3