Block hits and hidden blocks implemented
Now Mario can hit blocks! It looks and feels so amazing. I liked the way Math was used to animate it. It just used the Vector3.Lerp() for both going up and coming down when Mario hit the block with his head.
Block collision detection done using DotProd() defined in the Extensions script to check if Mario is moving in the upward direction to hit the block.
Hidden blocks are a piece of cake, just disable the sprite renderer as it gets enabled each time the HitBlock() function is called during a collision event. However, Mario can stand on the invisible block even when it is not rendered. Hopefully that would be fixed soon.