Coins in blocks and power up spawn implemented
Coin has 2 separate prefabs. One for when it is inside the block and the other for when it is placed on the scene.
Animation for coin bouncing off the block when hit is implemented (obviously using Math).
Prefabs for power ups have been implemented. When placed inside a block, they spawn with their own animation.
Reduced the size of the powerups to 0.925 uniformly for making them not clip through the edges of the mystery block.
I kinda don't like the coin block hit animation and I might fix that later in the end.
GameManager now updates the number of coins collected and adds an extra life when collected 100 coins.