Camera underground height adjustment implemented
The camera now checks if the exit position of the warp is less than 4 units under ground to set the boolean inUnderGround to true. Yup, its 4 units under ground and I hate to set it to a constant value like that instead of less than 0 but as per how my scene is constructed, the value is assigned accordingly.